Sticker Vers la Ville
Living in the city
You know you have to survive
Living in the city
You've got to keep the dream alive
Living in the city
Where everything is free
Living in the city
Can't you see?

Living in the city
You know you have to survive
Living in the city
You've got to keep the dream alive
Living in the city
Where everything is free
Living in the city
Can't you see?

So many different things to see
There's no time
So many different things to do
But there's no time

So many people all around
So many feelings to be found
Living in a city where no one lets you down

Living in the city
You know you have to survive
Living in the city
You've got to keep the dream alive
Living in the city
Where everything is free
Living in the city
Can't you see?

Living in the city, in the city
You know you have to survive
Living in the city
You've got to keep the dream alive
Living in the city
Where everything is free
Living in the city
Can't you see?

So many things that we've been through
Can you feel it?
So many things for us to do
But can you feel it?

So many faces all around
So many places to be found
Living in a city where nothing gets you down
Oh...

Living in the city
You know you have to survive
Living in the city
You've got to keep the dream alive
Living in the city
Where everything is free
Living in the city
Can't you see?

Living in the city
You know you have to survive
Living in the city
You've got to keep the dream alive
Living in the city
Where everything is free
Living in the city
Can't you see?

In the sky at night
The stars are shinin' bright
You can see the city lights '
cause I'm here to make it right

In the city...
Oh...
Yeah...
Oh...

Living in the city
You know you have to survive
Living in the city
You've got to keep the dream alive
Living in the city
Where everything is free
Living in the city
Can't you see?

Living in the city
You know you have to survive
Living in the city
You've got to keep the dream alive
Living in the city
Where everything is free
Living in the city
Can't you see?

Living in the city
Y'ou know you have to survive
Living in the city
You've got to keep the dream alive
Living in the city
Where everything is free
Living in the city
Can't you see?

In the city, in the city
Living in the city
In the city, in the city
Living in the city
Oh...
In the city
Living in the city
Living in the city
You know you have to survive
Living in the city
Oh, oh, oh...
Living in the city
Oh, oh, oh, oh, oh...
Living in the city
In the city, in the city
Living in the city In the city, in the city
Living in the city
Living in the city
Living in the city